split volume visual code
